Adds DockerHub login and debugs secrets
Adds a debug step to check DockerHub secrets and logs in to Docker Hub in the CI/CD workflow.

This ensures that the DockerHub credentials are correctly configured before attempting to push images, preventing potential deployment failures due to authentication issues.
It also sets the .NET SDK version to 8.0.x