Simplifies CI/CD branching strategy
Updates the CI/CD pipeline to focus on 'master' and 'develop' branches.

Removes 'main' branch from the trigger configuration for both push and pull request events, streamlining the build and deployment process.

Uses `pull_request_target` for pull requests to allow workflows to run in the context of the target branch.